requirements.txt
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
contourpy==1.0.7
cycler==0.11.0
filelock==3.9.0
Flask==2.2.3
Flask-Cors==3.0.10
fonttools==4.39.0
huggingface-hub==0.13.0
idna==3.4
importlib-metadata==6.0.0
importlib-resources==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.4
layoutlmv3 @ file:///Users/leelee24242424/developer/xaana/LayoutLm/unilm/layoutlmv3
MarkupSafe==2.1.2
matplotlib==3.7.1
numpy==1.23.4
opencv-python==4.5.5.62
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
pyparsing==3.0.9
pytesseract==0.3.10
python-dateutil==2.8.2
pytz==2022.7.1
PyYAML==6.0
regex==2022.10.31
requests==2.28.2
six==1.16.0
tokenizers==0.13.2
torch==1.13.0
torchvision==0.14.0
tqdm==4.65.0
transformers @ file:///Users/leelee24242424/Downloads/transformers-4.25.1
typing_extensions==4.5.0
urllib3==1.26.14
Werkzeug==2.2.3
zipp==3.15.0
